Moved methods to fsc to fix bug where Mayo cannot load numpy or pandas.
